# openapi.yaml
swagger: "2.0"
info:
  title: "Safe At Home"
  description: "Rest API for Michal Sela hackathon"
  version: "1.0.0"
# This field will be replaced by the deploy_api.sh script.
host: "YOUR-PROJECT-ID.appspot.com"
schemes:
  - "https"
definitions:
  Report:
    properties:
      token:
        type: string
      firstName:
        type: string
      LastName:
        type: string
    required:
      - token
      - firstName
      - LastName
  NewPerson:
    properties:
      token:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      id_num:
        type: string
    required:
      - token
      - firstName
      - LastName
      - id_num
paths:
  "/airportName":
    get:
      description: "Get the airport name for a given IATA code."
      operationId: "airportName"
      parameters:
        -
          name: iataCode
          in: query
          required: true
          type: string
      responses:
        200:
          description: "Success."
          schema:
            type: string
        400:
          description: "The IATA code is invalid or missing."
  "/searchPerson":
    get:
      description: "Get the information about a single person."
      operationId: "searchPerson"
      parameters:
        -
          name: firstName
          in: query
          required: true
          type: string
        -
          name: lastName
          in: query
          required: true
          type: string
      responses:
        200:
          description: "Success."
          schema:
            type: string
        400:
          description: "There was some error."
  "/users/list":
    get:
      description: "get users list"
      operationId: "get_users_list"
      responses:
        200:
          description: "Success."
          schema:
            type: string
        400:
          description: "There was some error."
  "/criminals/list":
    get:
      description: "get criminals list"
      operationId: "get_criminals_list"
      responses:
        200:
          description: "Success."
          schema:
            type: string
        400:
          description: "There was some error."
  "/person":
    get:
      description: "get person report"
      operationId: "get_persons_report"
      parameters:
        - name: firstName
          in: query
          required: true
          type: string
        - name: lastName
          in: query
          required: true
          type: string
        - name: token
          in: query
          required: true
          type: string
      responses:
        200:
          description: "Success."
          schema:
            type: string
        400:
          description: "There was some error."
  "/newPerson":
    post:
      description: "sign up - insert a new person"
      operationId: "insert_person"
      parameters:
        - in: body
          name: newPerson
          schema:
            $ref: '#/definitions/NewPerson'
      responses:
        200:
          description: "Success."
          schema:
            type: string
        400:
          description: "There was some error."
  "/report":
    post:
      description: "create a person report"
      operationId: "create_report"
      parameters:
        - in: body
          name: report
          schema:
            $ref: '#/definitions/Report'
      responses:
        200:
          description: "Success."
          schema:
            type: string
        400:
          description: "There was some error."